+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - **[UPDATE]** Update `TopBar/Menu` and `ItemsList` to support a single child item
+- **[BREAKING CHANGE]** Update `title` prop to `mainTitle` for `Hint`, `Profil`, `TripCard` and `QrCard`
 [...]
 
 # v53.0.0 (05/03/2021)